# CI: Speed up static checks by checking only changed files