--- a/.github/workflows/runner.yml
+++ b/.github/workflows/runner.yml
@@ -7,7 +7,7 @@ concurrency:
 
 jobs:
   static-checks:
-    name: 📊 Static
+    name: 📊 Static checks
     uses: ./.github/workflows/static_checks.yml
 
   android-build: