Our static checks used to take close to 8m to run, as many of the checks run on most files in the codebase each time, and are relatively slow bash or python scripts working file by file.

This PR optimizes the workflow by changing the scripts to (optionally) operate on a list of files, and we pass in the list of files actually changed in the PR.

We disable the dotnet format check if no `modules/mono` files are changed, as it's very slow and couldn't be optimized the same way as others. And we do the same for the JS for an extra 15 s speedup.

What's left as potential to optimize would be the dependencies install. Maybe the pip stuff can be cached too (see [actions/setup-python](https://github.com/actions/setup-python#caching-packages-dependencies)).

Before:
![image](https://user-images.githubusercontent.com/4701338/233152362-c87af42b-f0e7-40c3-aae0-1d203feabfef.png)

After:
![image](https://user-images.githubusercontent.com/4701338/233154638-14a18699-ea3b-4e0a-bdf0-bb1ca8efb3a2.png)

So we go from 7m 45s to ~1m 5s~ **50s**.