@Calinou 2023-04-19
Another possible optimization is to run each static check in parallel, and abort as soon as one of them reports an error.