@RedworkDE 2023-04-19
Possible edge case: Some huge PR with thousands of changed files could exceed the maximum size of the argument list for the scripts and fail that way. It might be better to pass the file with the changed files instead of the changed files directly.