@akien-mga 2023-04-20
> Possible edge case: Some huge PR with thousands of changed files could exceed the maximum size of the argument list for the scripts and fail that way. It might be better to pass the file with the changed files instead of the changed files directly.

Updated to do this.

> What's left as potential to optimize would be the dependencies install. Maybe the pip stuff can be cached too (see [actions/setup-python](https://github.com/actions/setup-python#caching-packages-dependencies)).

I reduced the setup time for dependencies by moving the static checks to Ubuntu 22.04, which provides clang-format 14 which is recent enough, so we don't need to install it from the LLVM repo anymore.

It's down from ca. 30 s to 15 s in average. Though sometimes it still takes longer if for whatever reason the Ubuntu or pip repos are slow. We could attempt to cache pip packages, but we should see if the `setup-python` action's own setup wouldn't be slower than the time we can save with caching.

The codespell action also sets up an older version of Python internally, but I'm not sure we can control that and share the python versions.