@akien-mga 2023-04-20
> I reduced the setup time for dependencies by moving the static checks to Ubuntu 22.04, which provides clang-format 14 which is recent enough, so we don't need to install it from the LLVM repo anymore.
> 
> It's down from ca. 30 s to 15 s in average. Though sometimes it still takes longer if for whatever reason the Ubuntu or pip repos are slow, like this: ![image](https://user-images.githubusercontent.com/4701338/233312133-d0df7431-7852-4990-be40-7aaf10229a9a.png)

To reduce this further, I'm adding https://github.com/marketplace/actions/cache-apt-packages to cache the APT packages. This seems to reduce it more consistently to around 9-12 s for the current packages we have setup, which is a bit faster than a "good" APT run, but also prevents us from getting into issues with either Ubuntu or Microsoft repositories being wonky.

> We could attempt to cache pip packages, but we should see if the `setup-python` action's own setup wouldn't be slower than the time we can save with caching.

Tried that, it works but it's indeed slower.

Installing our current pip dependencies straight with the pip from the Ubuntu 22.04 image takes ~6-7 s, while setting up python again with pip cache + restoring dependencies from cache seems to take between 12 and 18 s in my tests.

---

Final result, I think I'm stopping here :)

**54s for a trivial build.**

Should scale well with more files touched in bigger PRs, linearly to a max a bit below what it used to be (maybe 7m if changing all files in the codebase).

![image](https://user-images.githubusercontent.com/4701338/233323998-212c72e7-dc2f-478c-b4c0-b9778c988ebb.png)