@akien-mga 2023-04-20
Tweaked the job name as the previous list of scripts was incomplete, so I simplified it:
![image](https://user-images.githubusercontent.com/4701338/233327592-12ee2903-3b86-4c3c-9267-4a7ba9e592bb.png)

And also added conditions to skip Python checks when no Python files are changed, shaves off a few seconds more.

I'm done force push tweaks :)