@YuriSizov 2023-04-26
Cherry-picked for 4.0.3.